# netbox-backend/api/netbox_go/v1/extrasConfigcontextSiteGroups_router.pb.go
// Code generated by https://github.com/go-dev-frame/sponge, DO NOT EDIT.

package v1

import (
	"context"
	"errors"
	"strings"

	"github.com/gin-gonic/gin"
	"go.uber.org/zap"

	"github.com/go-dev-frame/sponge/pkg/errcode"
	"github.com/go-dev-frame/sponge/pkg/gin/middleware"
)

type ExtrasConfigcontextSiteGroupsLogicer interface {
	Create(ctx context.Context, req *CreateExtrasConfigcontextSiteGroupsRequest) (*CreateExtrasConfigcontextSiteGroupsReply, error)
	DeleteByID(ctx context.Context, req *DeleteExtrasConfigcontextSiteGroupsByIDRequest) (*DeleteExtrasConfigcontextSiteGroupsByIDReply, error)
	UpdateByID(ctx context.Context, req *UpdateExtrasConfigcontextSiteGroupsByIDRequest) (*UpdateExtrasConfigcontextSiteGroupsByIDReply, error)
	GetByID(ctx context.Context, req *GetExtrasConfigcontextSiteGroupsByIDRequest) (*GetExtrasConfigcontextSiteGroupsByIDReply, error)
	List(ctx context.Context, req *ListExtrasConfigcontextSiteGroupsRequest) (*ListExtrasConfigcontextSiteGroupsReply, error)
	DeleteByIDs(ctx context.Context, req *DeleteExtrasConfigcontextSiteGroupsByIDsRequest) (*DeleteExtrasConfigcontextSiteGroupsByIDsReply, error)
	GetByCondition(ctx context.Context, req *GetExtrasConfigcontextSiteGroupsByConditionRequest) (*GetExtrasConfigcontextSiteGroupsByConditionReply, error)
	ListByIDs(ctx context.Context, req *ListExtrasConfigcontextSiteGroupsByIDsRequest) (*ListExtrasConfigcontextSiteGroupsByIDsReply, error)
	ListByLastID(ctx context.Context, req *ListExtrasConfigcontextSiteGroupsByLastIDRequest) (*ListExtrasConfigcontextSiteGroupsByLastIDReply, error)
}

type ExtrasConfigcontextSiteGroupsOption func(*extrasConfigcontextSiteGroupsOptions)

type extrasConfigcontextSiteGroupsOptions struct {
	isFromRPC  bool
	responser  errcode.Responser
	zapLog     *zap.Logger
	httpErrors []*errcode.Error
	rpcStatus  []*errcode.RPCStatus
	wrapCtxFn  func(c *gin.Context) context.Context
}

func (o *extrasConfigcontextSiteGroupsOptions) apply(opts ...ExtrasConfigcontextSiteGroupsOption) {
	for _, opt := range opts {
		opt(o)
	}
}

func WithExtrasConfigcontextSiteGroupsHTTPResponse() ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.isFromRPC = false
	}
}

func WithExtrasConfigcontextSiteGroupsRPCResponse() ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.isFromRPC = true
	}
}

func WithExtrasConfigcontextSiteGroupsResponser(responser errcode.Responser) ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.responser = responser
	}
}

func WithExtrasConfigcontextSiteGroupsLogger(zapLog *zap.Logger) ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.zapLog = zapLog
	}
}

func WithExtrasConfigcontextSiteGroupsErrorToHTTPCode(e ...*errcode.Error) ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.httpErrors = e
	}
}

func WithExtrasConfigcontextSiteGroupsRPCStatusToHTTPCode(s ...*errcode.RPCStatus) ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.rpcStatus = s
	}
}

func WithExtrasConfigcontextSiteGroupsWrapCtx(wrapCtxFn func(c *gin.Context) context.Context) ExtrasConfigcontextSiteGroupsOption {
	return func(o *extrasConfigcontextSiteGroupsOptions) {
		o.wrapCtxFn = wrapCtxFn
	}
}

func RegisterExtrasConfigcontextSiteGroupsRouter(
	iRouter gin.IRouter,
	groupPathMiddlewares map[string][]gin.HandlerFunc,
	singlePathMiddlewares map[string][]gin.HandlerFunc,
	iLogic ExtrasConfigcontextSiteGroupsLogicer,
	opts ...ExtrasConfigcontextSiteGroupsOption) {

	o := &extrasConfigcontextSiteGroupsOptions{}
	o.apply(opts...)

	if o.responser == nil {
		o.responser = errcode.NewResponser(o.isFromRPC, o.httpErrors, o.rpcStatus)
	}
	if o.zapLog == nil {
		o.zapLog, _ = zap.NewProduction()
	}

	r := &extrasConfigcontextSiteGroupsRouter{
		iRouter:               iRouter,
		groupPathMiddlewares:  groupPathMiddlewares,
		singlePathMiddlewares: singlePathMiddlewares,
		iLogic:                iLogic,
		iResponse:             o.responser,
		zapLog:                o.zapLog,
		wrapCtxFn:             o.wrapCtxFn,
	}
	r.register()
}

type extrasConfigcontextSiteGroupsRouter struct {
	iRouter               gin.IRouter
	groupPathMiddlewares  map[string][]gin.HandlerFunc
	singlePathMiddlewares map[string][]gin.HandlerFunc
	iLogic                ExtrasConfigcontextSiteGroupsLogicer
	iResponse             errcode.Responser
	zapLog                *zap.Logger
	wrapCtxFn             func(c *gin.Context) context.Context
}

func (r *extrasConfigcontextSiteGroupsRouter) register() {
	r.iRouter.Handle("POST", "/api/extras/extrasConfigcontextSiteGroups", r.withMiddleware("POST", "/api/extras/extrasConfigcontextSiteGroups", r.Create_174)...)
	r.iRouter.Handle("DELETE", "/api/extras/extrasConfigcontextSiteGroups/:id", r.withMiddleware("DELETE", "/api/extras/extrasConfigcontextSiteGroups/:id", r.DeleteByID_170)...)
	r.iRouter.Handle("PUT", "/api/extras/extrasConfigcontextSiteGroups/:id", r.withMiddleware("PUT", "/api/extras/extrasConfigcontextSiteGroups/:id", r.UpdateByID_170)...)
	r.iRouter.Handle("GET", "/api/extras/extrasConfigcontextSiteGroups/:id", r.withMiddleware("GET", "/api/extras/extrasConfigcontextSiteGroups/:id", r.GetByID_170)...)
	r.iRouter.Handle("POST", "/api/extras/extrasConfigcontextSiteGroups/list", r.withMiddleware("POST", "/api/extras/extrasConfigcontextSiteGroups/list", r.List_174)...)
	r.iRouter.Handle("POST", "/api/extras/extrasConfigcontextSiteGroups/delete/ids", r.withMiddleware("POST", "/api/extras/extrasConfigcontextSiteGroups/delete/ids", r.DeleteByIDs_170)...)
	r.iRouter.Handle("POST", "/api/extras/extrasConfigcontextSiteGroups/condition", r.withMiddleware("POST", "/api/extras/extrasConfigcontextSiteGroups/condition", r.GetByCondition_174)...)
	r.iRouter.Handle("POST", "/api/extras/extrasConfigcontextSiteGroups/list/ids", r.withMiddleware("POST", "/api/extras/extrasConfigcontextSiteGroups/list/ids", r.ListByIDs_170)...)
	r.iRouter.Handle("GET", "/api/extras/extrasConfigcontextSiteGroups", r.withMiddleware("GET", "/api/extras/extrasConfigcontextSiteGroups", r.ListByLastID_170)...)

}

func (r *extrasConfigcontextSiteGroupsRouter) withMiddleware(method string, path string, fn gin.HandlerFunc) []gin.HandlerFunc {
	handlerFns := []gin.HandlerFunc{}

	// determine if a route group is hit or miss, left prefix rule
	for groupPath, fns := range r.groupPathMiddlewares {
		if groupPath == "" || groupPath == "/" {
			handlerFns = append(handlerFns, fns...)
			continue
		}
		size := len(groupPath)
		if len(path) < size {
			continue
		}
		if groupPath == path[:size] {
			handlerFns = append(handlerFns, fns...)
		}
	}

	// determine if a single route has been hit
	key := strings.ToUpper(method) + "->" + path
	if fns, ok := r.singlePathMiddlewares[key]; ok {
		handlerFns = append(handlerFns, fns...)
	}

	return append(handlerFns, fn)
}

func (r *extrasConfigcontextSiteGroupsRouter) Create_174(c *gin.Context) {
	req := &CreateExtrasConfigcontextSiteGroupsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.Create(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) DeleteByID_170(c *gin.Context) {
	req := &DeleteExtrasConfigcontextSiteGroupsByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) UpdateByID_170(c *gin.Context) {
	req := &UpdateExtrasConfigcontextSiteGroupsByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.UpdateByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) GetByID_170(c *gin.Context) {
	req := &GetExtrasConfigcontextSiteGroupsByIDRequest{}
	var err error

	if err = c.ShouldBindUri(req); err != nil {
		r.zapLog.Warn("ShouldBindUri error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) List_174(c *gin.Context) {
	req := &ListExtrasConfigcontextSiteGroupsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.List(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) DeleteByIDs_170(c *gin.Context) {
	req := &DeleteExtrasConfigcontextSiteGroupsByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.DeleteByIDs(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) GetByCondition_174(c *gin.Context) {
	req := &GetExtrasConfigcontextSiteGroupsByConditionRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.GetByCondition(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) ListByIDs_170(c *gin.Context) {
	req := &ListExtrasConfigcontextSiteGroupsByIDsRequest{}
	var err error

	if err = c.ShouldBindJSON(req); err != nil {
		r.zapLog.Warn("ShouldBindJSON error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByIDs(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}

func (r *extrasConfigcontextSiteGroupsRouter) ListByLastID_170(c *gin.Context) {
	req := &ListExtrasConfigcontextSiteGroupsByLastIDRequest{}
	var err error

	if err = c.ShouldBindQuery(req); err != nil {
		r.zapLog.Warn("ShouldBindQuery error", zap.Error(err), middleware.GCtxRequestIDField(c))
		r.iResponse.ParamError(c, err)
		return
	}

	var ctx context.Context
	if r.wrapCtxFn != nil {
		ctx = r.wrapCtxFn(c)
	} else {
		ctx = middleware.WrapCtx(c)
	}

	out, err := r.iLogic.ListByLastID(ctx, req)
	if err != nil {
		if errors.Is(err, errcode.SkipResponse) {
			return
		}
		r.iResponse.Error(c, err)
		return
	}

	r.iResponse.Success(c, out)
}